+++ documentation/plots/secondary_locations.py
@@ -9,8 +9,11 @@ def configure(context):
     context.stage("analysis.synthesis.mode_distances")
     context.stage("analysis.reference.hts.mode_distances")
 
+    context.config("hts")
+
 def execute(context):
     plotting.setup()
+    hts_name = context.config("hts")
 
     # PLOT: Input distributions
     distributions = context.stage("synthesis.population.spatial.secondary.distance_distributions")
@@ -76,11 +79,11 @@ def execute(context):
         plt.subplot(1, 2, index + 1)
 
         mode_reference = reference_data[mode]
-        plt.plot(mode_reference["values"] * 1e-3, mode_reference["cdf"], linestyle = '--', color = "k", linewidth = 1.0, label = "EGT")
+        plt.plot(mode_reference["values"] * 1e-3, mode_reference["cdf"], linestyle = '--', color = "k", linewidth = 1.0, label = "HTS")
 
         df_mode = df_synthetic[df_synthetic["mode"] == mode]
-        plt.fill_betweenx(df_mode["cdf"], df_mode["q5"]* 1e-3, df_mode["q95"] * 1e-3, linewidth = 0.0, color = plotting.COLORS["egt"], alpha = 0.25, label = "90% Conf.")
-        plt.plot(df_mode["mean"] * 1e-3, df_mode["cdf"], color = plotting.COLORS["egt"], linewidth = 1.0, label = "Synthetic")
+        plt.fill_betweenx(df_mode["cdf"], df_mode["q5"]* 1e-3, df_mode["q95"] * 1e-3, linewidth = 0.0, color = plotting.COLORS[hts_name], alpha = 0.25, label = "90% Conf.")
+        plt.plot(df_mode["mean"] * 1e-3, df_mode["cdf"], color = plotting.COLORS[hts_name], linewidth = 1.0, label = "Synthetic")
 
         plt.xlim([0, limits[mode] * 1e-3])
         plt.ylim([0, 1])