+++ synthesis/population/spatial/secondary/components.py
@@ -10,7 +10,7 @@ class CustomDistanceSampler(rda.FeasibleDistanceSampler):
         self.distributions = distributions
 
     def sample_distances(self, problem):
-        distances = np.zeros((problem["size"] + 1))
+        distances = np.zeros((len(problem["modes"])))
 
         for index, (mode, travel_time) in enumerate(zip(problem["modes"], problem["travel_times"])):
             mode_distribution = self.distributions[mode]
@@ -24,7 +24,7 @@ class CustomDistanceSampler(rda.FeasibleDistanceSampler):
 
         return distances
 
-class CustomDiscretizationSolver(rda.DiscretizationSolver):
+class CandidateIndex:
     def __init__(self, data):
         self.data = data
         self.indices = {}
@@ -33,16 +33,47 @@ class CustomDiscretizationSolver(rda.DiscretizationSolver):
             print("Constructing spatial index for %s ..." % purpose)
             self.indices[purpose] = sklearn.neighbors.KDTree(data["locations"])
 
+    def query(self, purpose, location):
+        index = self.indices[purpose].query(location.reshape(1, -1), return_distance = False)[0][0]
+        identifier = self.data[purpose]["identifiers"][index]
+        location = self.data[purpose]["locations"][index]
+        return identifier, location
+
+    def sample(self, purpose, random):
+        index = random.randint(0, len(self.data[purpose]["locations"]))
+        identifier = self.data[purpose]["identifiers"][index]
+        location = self.data[purpose]["locations"][index]
+        return identifier, location
+
+class CustomDiscretizationSolver(rda.DiscretizationSolver):
+    def __init__(self, index):
+        self.index = index
+
     def solve(self, problem, locations):
         discretized_locations = []
         discretized_identifiers = []
 
         for location, purpose in zip(locations, problem["purposes"]):
-            index = self.indices[purpose].query(location.reshape(1, -1), return_distance = False)[0][0]
+            identifier, location = self.index.query(purpose, location.reshape(1, -1))
 
-            discretized_identifiers.append(self.data[purpose]["identifiers"][index])
-            discretized_locations.append(self.data[purpose]["locations"][index])
+            discretized_identifiers.append(identifier)
+            discretized_locations.append(location)
+
+        assert len(discretized_locations) == problem["size"]
 
         return dict(
             valid = True, locations = np.vstack(discretized_locations), identifiers = discretized_identifiers
         )
+
+class CustomFreeChainSolver(rda.RelaxationSolver):
+    def __init__(self, random, index):
+        self.random = random
+        self.index = index
+
+    def solve(self, problem, distances):
+        identifier, anchor = self.index.sample(problem["purposes"][0], self.random)
+        locations = rda.sample_tail(self.random, anchor, distances)
+        locations = np.vstack((anchor, locations))
+
+        assert len(locations) == len(distances) + 1
+        return dict(valid = True, locations = locations)