+++ synthesis/population/spatial/secondary/problems.py
@@ -58,7 +58,7 @@ def find_assignment_problems(df, df_locations):
             problem["purposes"] = problem["purposes"][:-1]
 
         else:
-            raise RuntimeError("The presented 'problem' is neither a chain nor a tail")
+            pass # Neither chain nor tail
 
         # Define size
         problem["size"] = len(problem["purposes"])
@@ -81,5 +81,10 @@ def find_assignment_problems(df, df_locations):
         if destination_purpose in FIXED_PURPOSES:
             problem["destination"] = current_location[LOCATION_FIELDS.index(destination_purpose)] # Shapely POINT
             problem["destination"] = np.array([[problem["destination"].x, problem["destination"].y]])
+
+        if problem["origin"] is None:
+            problem["activity_index"] = problem["trip_index"]
+        else:
+            problem["activity_index"] = problem["trip_index"] + 1
 
         yield problem