+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 **Under development**
 
+- Update to `eqasim-java:1.2.0` to support tails and "free" activity chains
 - Allow for activity chains that do not start and end at home
 - Improve handling of education attribute in ENTD