+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 **Under development**
 
+- Bugfix: Make sure df_trips are sorted properly in `synthesis.population.trips`
 - Bugfix: Do not execute "urban" attribute imputation twice
 - Bugfix: Do not consider *inactive* enterprises from SIRENE
 - Update analysis scripts