+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 **Under development**
 
+- By default, load SIRENE directly from `zip` file instead of `csv`
 - Bugfix: Make sure df_trips are sorted properly in `synthesis.population.trips`
 - Bugfix: Do not execute "urban" attribute imputation twice
 - Bugfix: Do not consider *inactive* enterprises from SIRENE