+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 **Under development**
 
+- No changes yet
+
+**1.2.0**
+
 - Update code and data to BPE 2019 (verison for 2018 is not available anymore)
 - Add additional spatial standard output: `homes.gpkg` and `commutes.gpkg`
 - Updated documentation for BD-TOPO